+++ src/bzfs/bzfs.cxx
@@ -3633,7 +3633,7 @@ int main(int argc, char **argv)
 
 #endif /* defined(_WIN32) */
 
-  bzfsrand(time(0));
+  bzfsrand((unsigned int)time(0));
 
   Flags::init();
